requirements.txt
anyio==4.0.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
notion-client==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
